+++ test/e2e/features/cat-ce/10-pass_with_no_faults.feature
@@ -2,16 +2,16 @@
 Feature: A Driving Examiner Completes a Passed User Journey in Category CE
 
    Scenario: Examiner completes a passed test with no faults
-      Given I am logged in as "mobexaminer5" and I have a test for "Mr Right Ford"
-      When I start the test for "Mr Right Ford"
+      Given I am logged in as "mobexaminer5" and I have a test for "Ms Diaz Barber"
+      When I start the test for "Ms Diaz Barber"
       And the candidate completes the declaration page
       And the candidate confirms their declaration
-      Then I should see the "Declaration - Right Ford" page
+      Then I should see the "Declaration - Diaz Barber" page
       And the candidate enters a new email address
       And I proceed to the car
-      Then I should see the "Right Ford" page
+      Then I should see the "Diaz Barber" page
       And I complete the waiting room to car page
-      Then I should see the "Test report - Right Ford" page
+      Then I should see the "Test report - Diaz Barber" page
       When I open the reversing diagram
       Then I should see the reversing diagram modal
       And I close the reversing diagram modal
@@ -20,7 +20,7 @@ Feature: A Driving Examiner Completes a Passed User Journey in Category CE
       And I continue to debrief
       Then I should see the Debrief page with outcome "Passed"
       When I end the debrief
-      Then I should see the "Test debrief - Right Ford" page
+      Then I should see the "Test debrief - Diaz Barber" page
       And  I select the code 78 option
       And I complete the pass details
       And I complete the health declaration
@@ -31,4 +31,4 @@ Feature: A Driving Examiner Completes a Passed User Journey in Category CE
       When I complete the office write up
       And I upload the test
       Then I should see the "Journal" page
-      And the test result for "Mr Right Ford" is "1"
+      And the test result for "Ms Diaz Barber" is "1"
