+++ test/e2e/step-definitions/debrief-steps.ts
@@ -1,7 +1,6 @@
 import { Then, When, Before } from 'cucumber';
-import { getElement, clickElement } from './generic-steps';
 import { by } from 'protractor';
-import { textFieldInputViaNativeMode } from '../../helpers/helpers';
+import { textFieldInputViaNativeMode, getElement, clickElement } from '../../helpers/helpers';
 
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
