+++ test/e2e/step-definitions/postdebriefholding-steps.ts
@@ -1,6 +1,6 @@
 import { Then, When, Before } from 'cucumber';
-import { getElement, clickElement } from './generic-steps';
 import { browser, by, ExpectedConditions } from 'protractor';
+import { getElement, clickElement } from '../../helpers/helpers';
 
 this.testCategory = 'b';
